Moved Editor logic to Editor folder with Assembly ref
