+++ Assets/Game/Scripts/Emote/EyePose/Editor.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 31b662eb163342589762284b6508e841
+timeCreated: 1737832051
