+++ Assets/Game/Scripts/Emote/EyePose/Editor/EditorRef.asmref
@@ -1,0 +1,3 @@
+{
+    "reference": "GUID:75267a816038d684197863a57ca53466"
+}
