+++ Assets/Game/Scripts/Emote/EyePose/Editor/EditorRef.asmref.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 67d10244e4e74a8da14e5a55ce60f080
+timeCreated: 1737832172
