+++ Assets/Game/Scripts/Emote/EyePose/Editor/EyePoseCreatorWindow.cs.meta
