+++ Assets/Game/Scripts/Emote/EyePose/Editor/EyePoseEditor.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 14dba4ec309b4a7fae9fba0496faf8ce
+timeCreated: 1737832288
